pyproject.toml
[project]
name = "test-render"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.133.0",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
